% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfuns.r
\name{avg.tae.normal}
\alias{avg.tae.normal}
\title{avg.tae.normal}
\usage{
avg.tae.normal(kk, n, m)
}
\arguments{
\item{kk}{repeat number}

\item{n}{nrow}

\item{m}{ncol}
}
\value{
averaged tae
}
\description{
avg.tae.normal
}
