% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfuns.r
\name{avg.tae.simu}
\alias{avg.tae.simu}
\title{avg.tae.simu}
\usage{
avg.tae.simu(kk, n, m, T)
}
\arguments{
\item{kk}{repeat number}

\item{n}{nrow}

\item{m}{ncol}

\item{T}{repeating time}
}
\value{
averaged tae
}
\description{
avg.tae.simu
}
