% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfuns.r
\name{l.vec.compute}
\alias{l.vec.compute}
\title{l.vec}
\usage{
l.vec.compute(k, cn.vec, m)
}
\arguments{
\item{k}{input by given functions}

\item{cn.vec}{calculated by given fucntions}

\item{m}{dimensions}
}
\value{
A vector, which will be used in other functions
}
\description{
l.vec
}
