% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfuns.r
\name{pmd.by.demands}
\alias{pmd.by.demands}
\title{pmd.by.demands}
\usage{
pmd.by.demands(x_vec, pp, t = 1000)
}
\arguments{
\item{x_vec}{result vec input by user}

\item{pp}{The pp is a probability matrix which will be input by user}

\item{t}{simulation repeat time}
}
\value{
The probability mass of PMD
}
\description{
pmd.by.demands
}
