% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfuns.r
\name{pmd}
\alias{pmd}
\title{probability density of Poisson multinomial distribution}
\usage{
pmd(pp, method = "DFT-CF", vec = c(0, 0, 0, 0, 0), t = 100)
}
\arguments{
\item{pp}{The pp is a probability matrix which will be input by user}

\item{method}{method selected by user to compute the probability mass}

\item{vec}{result vec input by user}

\item{t}{simulation repeat time}
}
\value{
The probability mass of PMD
}
\description{
probability density of Poisson multinomial distribution
}
