% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfuns.r
\name{ppmn}
\alias{ppmn}
\title{ppmn}
\usage{
ppmn(pp, x, method = "DFT-CF", t = 1000)
}
\arguments{
\item{pp}{input matrix of probabilities}

\item{x}{input vector}

\item{method}{method}

\item{t}{repeating time}
}
\value{
prob
}
\description{
ppmn
}
